Bug fix.
Description: Multiple copies of Copyright appeared in Makefile.in.  This was
due to automake copying the copyright right in the included files such as
config/commence.am.

Solution: Automake treats double hashes as comments and does not copy them
to Makefile.in.  Changed all the copyright notices in config/*.am to use
double hashes for the Copyright right notice.

Tested: kagiso via bin/reconfigure.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@13365 dab4d1a6-ed17-0410-a064-d1ae371a2980