Purpose: Code cleanup
Description:
Changed a macro in H5FDwindows.c from IO_BUF_SIZE to a more specific WINDOWS_MAX_BUF, and moved it to our H5pubconf.h, where users can customize it to their own preference.  

This value is needed because of a bug in the Windows _write function, it is unsafe to write buffers larger than 2GB-1 bytes in one write.

Tested:
VS2005 on WinXP.

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@13873 dab4d1a6-ed17-0410-a064-d1ae371a2980