Purpose: Remove obsolete Windows files
Description:
Previously, there was a very intricate process for converting VS6 project files to VS.NET format for testing on the command-line.  Now that we maintain VS.NET project files, this system is no longer neccessary.  The same builds are now carried out in hdf5build.BAT, so the previous files are obsolete.

Removing them, and updating the MANIFEST

Tested:
WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@13960 dab4d1a6-ed17-0410-a064-d1ae371a2980