Purpose: Update test directories in MinGW documentation
Description:
There are many tests that are unsupported on MinGW because of the way it handles certain syntax in shell scripts (particularly '/').  As a result, we list a few tests that users must comment out in various Makefile.in's.  This checkin updates the location of h5stat-- it was formerly in ./tools/misc, but has been moved it ./tools/h5stat.

Tested:
MinGW on WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@14316 dab4d1a6-ed17-0410-a064-d1ae371a2980