Purpose: Fix typo in our configure CYGWIN patch
Description:
Previously, we created a patch for Cygwin to skip checking for the timezone variable.  We made this change in both the trunk and hdf5_1_8 branch.  However, in the branch version, we seemed to have used wrong quotes in our check ('uname' rather than `uname`).  This corrects the typo

Tested:
Cygwin on WinXP


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@14723 dab4d1a6-ed17-0410-a064-d1ae371a2980