Purpose: Fix error message when with Windows pthreads
Description:
When building HDF5 with thread safety on Windows, the err_compat test was failing because our output was being parsed incorrectly.  Rather than having a thread number in the error stack trace, there will be the string "some thread: no way to know the thread (IDs): from pthread on windows:".  This checkin now takes this into account, and modifies the output accordingly.

Tested:
VS2005 on WinXP with Pthreads

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@14855 dab4d1a6-ed17-0410-a064-d1ae371a2980