Purpose: Add Windows fixes for new Fortran code.
Description:
There were a number of small tweaks we needed to make to add the new fortran_1_8 code on Windows.  We create new project files, add new source to them, add the test to our test suite, and fix a few typos in the Windows-specific source code.

Tested:
VS2005 on WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@14975 dab4d1a6-ed17-0410-a064-d1ae371a2980