Purpose: bug fix.
Description:
Removed a few debug print statments that was left in by mistake.
Removed initialization for macro system INSTRUMENT to match how it
behaves before.

Tested platforms:
h5committested (kagiso, smirom and linew).  It failed in fortran
test in smirom but I believed that was not caused by my changes
in configure.in.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@15017 dab4d1a6-ed17-0410-a064-d1ae371a2980