Purpose: Add testhdf5_1_8fortran[dll] Windows projects as dependencies to all_fortran
Description:
The projects testhdf5_1_8fortran[dll] which test the new fortran features in HDF5 1.8 were not registered as dependencies to the "all_fortran" pseudo-project.  As a result, the projects would not be built if you simply build all_fortran in Visual Studio, rather than "Build Solution".  This is a very minor cleanup, but is neccessary to be consistent in our projects.

Tested:
VS2005 on WinXP x64

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@15153 dab4d1a6-ed17-0410-a064-d1ae371a2980