Purpose: Add new Fortran source file
Description:
There is a new Fortran file, H5_DBLE_InterfaceInclude.f90, which needs to be added to Windows.  There is also another version, *InterfaceExclude.f90, which is for users that build with the -r8 flag.  However, we don't include this on Windows, as it is not currently a common option.  If we get requests from users, we can add documentation for using this alternative interface.

Tested:
VS2005 w/ IVF9.1 on WinXP

git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@15546 dab4d1a6-ed17-0410-a064-d1ae371a2980