Bug 710: partial fix.
Temporary removed b2 from the --enable-debug=all list so that this option
will pass. "configure --help" lists the --enable-debug=all and users tend
to use it and end up with compile failures in the b2 debug code.  Since this
bug will not be fixed soon, this is a temporary fix to let the configure and
build work.

Tested:
h5committest plus "--enable-debu=all" test in kagiso.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@15637 dab4d1a6-ed17-0410-a064-d1ae371a2980