#1449 Change set_extent test to use standard standalone test framework
Call h5_fixname (with an array of test filenames) for generating the  
filename to create and then call h5_cleanup() when the tests pass, to delete the files  
created and close the FAPL from h5_fileaccess(). 

Defined a macro 
#define TESTING2(WHAT)	{printf(" Testing %-62s",WHAT); fflush(stdout);}
Similar to TESTING, except that it has a initial indentation space.
The effect is for nested loop tests

Testing with old file format:
 Testing with fill value, no compression                                PASSED

tested: windows, linux






git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@16419 dab4d1a6-ed17-0410-a064-d1ae371a2980