Removed the erroneous test of using v16 API. The error started out as an erroneous test of using v16 API via source code. (See bug 1264).  It was implemented as:
#include "hdf5.h"
#define H5_USE_16_API

That was incorrect and generated warnings in Linux but error in AIX
for redefining macro H5_USE_16_API.

I corrected it yesterday to pass AIX by moving H5_USE_16_API up as:

#define H5_USE_16_API
#include "hdf5.h"

This is the proper way to turn on v16API and it passes all normal
tests.  I knew it would not work if somehow v16API is disabled
but hoping it can be fixed later.  Well, it broke last night
under --disable-deprecated-symbols.

Removed the #define H5_USE_16_API and let Raymond fix it after
the release.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@17693 dab4d1a6-ed17-0410-a064-d1ae371a2980