Bug 1917: post cleanup.
fseek64 was used to support large file access for the STDIO driver back in
version 1.2.2 in year 2000.  Some how it was not included in version 1.4.0.
Now, fseeko64 is used to support large file. There is no more need for fseek64
which is not a standard call. Removed its presence from configure and related
files.

Tested: jam for configure only.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@19205 dab4d1a6-ed17-0410-a064-d1ae371a2980