Bug 2019: h5perf or h5perf_serial is removed twice. Some systems like AIX do not like it as it may get into a racing condition when "gmake -j N" is used-- two "rm" commands were trying to remove the same file.
Solution:
Rearranged the Makefile.am so that h5perf or h5perf_serial are built only
once, that is through bin_PROGRAMS only.

Bug 2135: h5perf_serial is not installed in --enable-parallel build.
It was because in parallel build, h5perf_serial is not was not in the
bin_PROGRAMS list but it was still being built and tested.

Solution:
Rearranged the Makefile.am code so that h5perf_serial is installed too.
(It is easier to make cleaner code to install it than to build and test
h5perf_serial but not test it.)

Tested: Jam, both serial and parallel.
Did not run h5committest because the changes are in perform/ directory only
and I think Jam has given the changes in Makefile a complete test already.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@20019 dab4d1a6-ed17-0410-a064-d1ae371a2980