Fixes h5diff test failure on BE systems.
H5Dread() produces different output for invalid enum type values
on systems of different endiannes.  On BE systems, it always
emits -1.  On LE systems, the invalid value is emitted.  This
difference caused the h5diff test to fail since it tests to
see if different invalid enum values are reported as different
(on BE systems they are not since -1 == -1).

This fix removes the "invalid vs. different invalid" test
while we fix the library bug.  HDFFV-7527 will be reopened
in JIRA until the full test passes.

Tested on: jam, heiwa, linew


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@20756 dab4d1a6-ed17-0410-a064-d1ae371a2980