HDFFV-8264: Using F2003, build fails on Fedora with undefined reference to __h5r_MOD_h5rget_region_region_f
This turned out being an issue with configure. The reporter submitted a patch which fixed the fact that we should not be 
setting AM_FCFLAGS (an automake variable) with FFLAGS (a user variable). 

I removed this, and we now only set FFLAGS if the environment variable is set, otherwise we don't.

Tested: jam (gnu)



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@23410 dab4d1a6-ed17-0410-a064-d1ae371a2980