Bug fix: HDFFV-8271 Big.c error in address generated by random number.
Description:
There is a bug in the current big.c tests. In a nut-shell,
the writing part randomly picks 50 locations and write 4KB
from that location. Random location range is from byte 0
to end of dataset. There is the problem--if the random location
is greater than (end of dataset - 4KB), then error.

Solution:
Adjust the random number to be at least 4KB less than the end of the
dataset. That allows it write 4KB more from the generated starting
position.

Tested: h5comittest plus jam serial build and test.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@24202 dab4d1a6-ed17-0410-a064-d1ae371a2980