Merge of r26071 from features/autotools_rework
Removes the --with-default-vfd option from configure. This was basically
useless and only allowed selecting the stdio VFD. Since this is a demo
VFD and not really a production VFD, we decided to retire this option.

Fixes: HDFFV-9081

Tested on: h5committest

NOTE: platypus cmake fails, but this is not a new error.



git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@26272 dab4d1a6-ed17-0410-a064-d1ae371a2980