Merge of r26069, r26070, and r26072 from the features/autotools_rework branch.
Removed the configure option that allows selective disabling of individual
internal filters (fletcher32, shuffle, etc.).

This feature mucked up the code with a lot of #ifdefs, saved very little space,
and was not scalable to a general scheme for library size reduction. We've
decided to remove the feature while we investigate a more general scheme for
decreasing the library size.

Part of: HDFFV-9086

Tested on: h5committest


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@26329 dab4d1a6-ed17-0410-a064-d1ae371a2980