Bug fix: cmakehdf5 broken because zlib is no longer configure in by default. That broke the testings as some testfiles have zlib compressed datasets.
Added options control to enable the linking of zlib external libarary by
default and turn off the szip library linking as szip library may not be
avaiable. This matches the established settings.

Tested: run cmakehdf5 by hand in jam and platypus.
Also tested in wren but it failed in the testing stage.
Also tried "cmakehdf5 --script" in jam. It failed.


git-svn-id: https://svn.hdfgroup.uiuc.edu/hdf5/trunk@26712 dab4d1a6-ed17-0410-a064-d1ae371a2980